adding addPostForm from MyPosts
